[SUBSCRIPTION] notification fix

+++ main.py
@@ -73,6 +73,8 @@ r = 10
 def delete_spam(user, agency):
     global messages_to_be_deleted
     global bot
+    if len(messages_to_be_deleted[user][agency]) < 2:
+        return
     for message, chat in messages_to_be_deleted[user][agency][1:]:
         bot.deleteMessage(chat, message)
     messages_to_be_deleted[user][agency] = []
@@ -91,9 +93,9 @@ def send_notification(subscription):
     global oldAvailable
     global r
     subscribed_users = telegramDb.get_subscribed_users(subscription)
+    requests = 0
+    bad_requests = 0
     for user in subscribed_users:
-        if user['_id'] != 1190803139:
-            continue
         for agency in user['availableSpace']:
             TS = AllAgencies[agency]
             if TS.maxDelegationCap == 'unlimited':
@@ -103,8 +105,10 @@ def send_notification(subscription):
             if not agency in old_available_values:
                 old_available_values[agency] = 0
             if newAvailable != old_available_values[agency]:
-                check_and_notify(user['_id'], newAvailable, old_available_values[agency], TS.name)
+                requests += 1
+                bad_requests += check_and_notify(user['_id'], newAvailable, old_available_values[agency], TS.name)
                 old_available_values[agency] = newAvailable
+    print("\tbad requests", str(bad_requests) + "/" + str(requests))
 
 
 def check_and_notify(user_id, newAvailable, oldAvailable, name):
@@ -122,6 +126,9 @@ def check_and_notify(user_id, newAvailable, oldAvailable, name):
                     + str(user_id) + '&parse_mode=Markdown&text=' + bot_message
         response = requests.get(send_text)
         data = response.json()
+        if not data['ok']:
+            print(user_id, ":", data)
+            return 0
         message_id = data['result']['message_id']
         chat_id = data['result']['chat']['id']
         time.sleep(0.1)
@@ -135,8 +142,16 @@ def check_and_notify(user_id, newAvailable, oldAvailable, name):
         send_text = 'https://api.telegram.org/bot' + bot_token + '/sendMessage?chat_id=' \
                     + str(user_id) + '&parse_mode=Markdown&text=' + bot_message
         response = requests.get(send_text)
+        data = response.json()
+        if not data['ok']:
+            print(user_id, ":", data)
+            return 0
         delete_spam(user_id, name)
         time.sleep(0.1)
+    else:
+        print("\t error: user_id", user_id, "agency", name, "newAvailable", newAvailable, 'oldAvailable', oldAvailable)
+        return 0
+    return 1
 
 
 def main():
